upstream: dpdk: speculative fix for hangs in dpdk_log_read_ready
The dpdk plugin sets up a pipe to read dpdk log messages and
then write them to the vpp logging mechanism. Occasionally vpp
would call read() on the read side of the pipe when no data
were available, and vpp's main thread would hang because read()
blocked.

This change sets the read side of the pipe to non-blocking.
There is already code in place after the read() call to
handle read() returns when no data are available.

Change-Id: I5f397aca0bf9ba5c9f9332e9f6793eddc75e18df