upstream: ipsec: parse only sad_id when deleting SA
Parsing the the IPsec protocol fails if the value has been initialized
to zero (or, indeed, is anything other than 50 or 51).  The only value
required to delete an SA is the identifier (sad_id), so do not parse
any other fields for delete.

Add test to ensure that only the ID is required.

Type: improvement
Signed-off-by: Eric Kinzie <ekinzie@labn.net>